Move mapped types to separate package
