+++ fusion-endpoint/src/test/java/com/vaadin/fusion/generator/endpoints/AbstractEndpointGenerationTest.java
@@ -83,11 +83,11 @@ import com.vaadin.fusion.EndpointExposed;
 import com.vaadin.fusion.auth.CsrfChecker;
 import com.vaadin.fusion.auth.FusionAccessChecker;
 import com.vaadin.fusion.endpointransfermapper.EndpointTransferMapper;
-import com.vaadin.fusion.endpointransfermapper.PageableDTO;
 import com.vaadin.fusion.generator.OpenAPIObjectGenerator;
 import com.vaadin.fusion.generator.endpoints.complexhierarchymodel.GrandParentModel;
 import com.vaadin.fusion.generator.endpoints.complexhierarchymodel.Model;
 import com.vaadin.fusion.generator.endpoints.complexhierarchymodel.ParentModel;
+import com.vaadin.fusion.mappedtypes.PageableDTO;
 import com.vaadin.fusion.utils.TestUtils;
 
 import static com.vaadin.fusion.ExplicitNullableTypeChecker.isRequired;
